pyproject.toml
[tool.poetry]
name = "turtlebot"
version = "0.1.2"
description = "A bot that allows you to create and run turtle scripts in Discord"
authors = ["Patrick Brennan (AM2i9)"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.3.1"
"discord.py" = "^1.7.3"
"dislash.py" = "^1.4.6"

[tool.poetry.dev-dependencies]
taskipy = "^1.8.1"
black = "^21.7b0"
python-dotenv = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"